Fixes #514: use double instead of int for duraction

+++ src/Media.Plugin.Android/MediaPickerActivity.cs
@@ -49,7 +49,7 @@ namespace Plugin.Media
         private bool saveToAlbum;
         private string action;
 
-        private int seconds;
+        private double seconds;
         private long size;
         private VideoQuality quality;
 
@@ -66,7 +66,7 @@ namespace Plugin.Media
             outState.PutInt(ExtraId, id);
             outState.PutString(ExtraType, type);
             outState.PutString(ExtraAction, action);
-            outState.PutInt(MediaStore.ExtraDurationLimit, seconds);
+            outState.PutDouble(MediaStore.ExtraDurationLimit, seconds);
             outState.PutLong(MediaStore.ExtraSizeLimit, size);
             outState.PutInt(MediaStore.ExtraVideoQuality, (int)quality);
             outState.PutBoolean(ExtraSaveToAlbum, saveToAlbum);
@@ -114,9 +114,10 @@ namespace Plugin.Media
                 {
                     if (!isPhoto)
                     {
-                        seconds = b.GetInt(MediaStore.ExtraDurationLimit, 0);
-                        if (seconds != 0)
-                            pickIntent.PutExtra(MediaStore.ExtraDurationLimit, seconds);
+                        seconds = b.GetDouble(MediaStore.ExtraDurationLimit, 0);
+						if (seconds != 0)
+							pickIntent.PutExtra(MediaStore.ExtraDurationLimit, seconds);
+						
                         size = b.GetLong(MediaStore.ExtraSizeLimit, 0);
                         if (size != 0)
                         {
